converted theme store into class
